ml_project/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
anyio @ file:///tmp/build/80754af9/anyio_1644463572971/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569679365/work
arrow==1.2.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
binaryornot==0.4.4
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
bokeh==2.4.3
brotlipy==0.7.0
catboost==1.1.1
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cookiecutter==2.1.1
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.11.0
dask==2022.10.2
dataprep==0.4.5
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926439650/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
graphviz==0.20.1
hydra-core==1.2.0
idna @ file:///croot/idna_1666125576474/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648544546694/work
ipykernel @ file:///opt/conda/conda-bld/ipykernel_1662361798230/work
ipython @ file:///opt/conda/conda-bld/ipython_1657652213665/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.2.0
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonpath-ng==1.5.3
jsonschema @ file:///opt/conda/conda-bld/jsonschema_1663375472438/work
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-console @ file:///opt/conda/conda-bld/jupyter_console_1647002188872/work
jupyter-server @ file:///tmp/abs_b88b31b8-83b9-476d-a46d-e563c421f38fvsnyi1ur/croots/recipe/jupyter_server_1658754481507/work
jupyter_client @ file:///opt/conda/conda-bld/jupyter_client_1662504365333/work
jupyter_core @ file:///opt/conda/conda-bld/jupyter_core_1664917302524/work
jupyterlab @ file:///tmp/abs_12f3h01vmy/croots/recipe/jupyterlab_1658907535764/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
jupyterlab_server @ file:///opt/conda/conda-bld/jupyterlab_server_1664908231281/work
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.6.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
Metaphone==0.6
mistune @ file:///tmp/build/80754af9/mistune_1607364877025/work
multidict==6.0.2
nbclassic @ file:///opt/conda/conda-bld/nbclassic_1644943264176/work
nbclient @ file:///tmp/build/80754af9/nbclient_1650290509967/work
nbconvert @ file:///opt/conda/conda-bld/nbconvert_1649751911790/work
nbformat @ file:///opt/conda/conda-bld/nbformat_1663744952973/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847906199/work
nltk==3.7
notebook @ file:///tmp/abs_abf6xa6h6f/croots/recipe/notebook_1659083654985/work
numpy==1.23.4
omegaconf==2.2.3
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.5.1
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd==1.3.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
plotly==5.11.0
ply==3.11
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.2
pydot==1.4.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636110951836/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-crfsuite==0.9.8
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-slugify==6.1.2
python-stdnum==1.17
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
qtconsole @ file:///opt/conda/conda-bld/qtconsole_1662018252641/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1662014892439/work
rapidfuzz==2.13.0
regex==2021.11.10
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
sniffio @ file:///tmp/build/80754af9/sniffio_1614030464178/work
soupsieve @ file:///croot/soupsieve_1666296392845/work
SQLAlchemy==1.3.24
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tenacity==8.1.0
terminado @ file:///tmp/build/80754af9/terminado_1644322582718/work
testpath @ file:///opt/conda/conda-bld/testpath_1655908557405/work
text-unidecode==1.3
threadpoolctl==3.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz==0.12.0
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///croot/urllib3_1666298941550/work
varname==0.8.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///tmp/build/80754af9/websocket-client_1614803975924/work
Werkzeug==2.2.2
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1644992802045/work
wordcloud==1.8.2.2
yarl==1.8.1
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
